requirements.txt
o365==2.1.9
gspread==6.2.1
google-auth==2.48.0
python-dotenv==1.2.2
pandas==2.2.2
streamlit==1.32.0
altair==4.2.2